only create symbolic link if it doesn't already exist
`-L` checks if it's a symbolic link

+++ static/change-to-zfs-mount-generator.sh
@@ -60,7 +60,10 @@ mkdir -p /etc/zfs/zfs-list.cache
 # Enable ZFS Event Daemon(ZED) aka ZEDLET
 if [ -f /usr/lib/zfs-linux/zed.d/history_event-zfs-list-cacher.sh ]
 then
-    check_command ln -s /usr/lib/zfs-linux/zed.d/history_event-zfs-list-cacher.sh /etc/zfs/zed.d
+    if [ ! -L /etc/zfs/zed.d/history_event-zfs-list-cacher.sh ]
+    then
+        check_command ln -s /usr/lib/zfs-linux/zed.d/history_event-zfs-list-cacher.sh /etc/zfs/zed.d
+    fi
 else
     msg_box "/usr/lib/zfs-linux/zed.d/history_event-zfs-list-cacher.sh is missing, aborting!"
     exit 1
